Split server/client plugin packets
